Refactored Swagger setup to its own file
