+++ apps/api/src/main.ts
@@ -6,11 +6,12 @@
 import { Logger, ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { ExpressAdapter } from '@nestjs/platform-express';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import cookieParser from 'cookie-parser';
 import express from 'express';
 import * as functions from 'firebase-functions';
 import { GithubModule } from './modules/github/github.module';
+import { setupAxios } from './setup/axios';
+import { setupSwagger } from './setup/swagger';
 
 const server = express();
 
@@ -18,13 +19,6 @@ async function bootstrap() {
 	const app = await NestFactory.create(GithubModule, new ExpressAdapter(server), { cors: true });
 	const globalPrefix = 'api';
 	const port = process.env.PORT || 3000;
-	const documentConfig = new DocumentBuilder()
-		.setTitle('Nestjs Github Project')
-		.setDescription('API description')
-		.setVersion('1.0')
-		.addTag('nestjs/github')
-		.build();
-	const documentFactory = () => SwaggerModule.createDocument(app, documentConfig);
 
 	app.setGlobalPrefix(globalPrefix);
 	app.use(cookieParser());
@@ -33,11 +27,16 @@ async function bootstrap() {
 			transform: true,
 		}),
 	);
-	SwaggerModule.setup('api/doc', app, documentFactory);
+
+	// set up other services and tools running together with the app
+	setupSwagger(app);
+	setupAxios();
+
 	await app.listen(port);
 	Logger.log(`🚀 Application is running on: http://localhost:${port}/${globalPrefix}`);
 }
 
 bootstrap();
 
+// Firebase integration
 export const api = functions.https.onRequest(server);
